Update BallIntake
Change the ball intake to use 2 speed controllers.  One is
on the lower assembly and used for pickup.  The second is
on the upper assembly and is used for intake and shooting.

For BallIntake, run the lower motor on high speed and the
upper motor on low.  For shooting, run the upper motor high.